backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
librosa==0.10.1
numpy==1.24.3
anthropic==0.7.7
yt-dlp==2023.11.16
spotipy==2.23.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
soundfile==0.12.1
